# generated/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.25.0

package db

import (
	"database/sql/driver"
	"fmt"

	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type UpdateProtocol string

const (
	UpdateProtocolExpo     UpdateProtocol = "expo"
	UpdateProtocolCodepush UpdateProtocol = "codepush"
)

func (e *UpdateProtocol) Scan(src interface{}) error {
	switch s := src.(type) {
	case []byte:
		*e = UpdateProtocol(s)
	case string:
		*e = UpdateProtocol(s)
	default:
		return fmt.Errorf("unsupported scan type for UpdateProtocol: %T", src)
	}
	return nil
}

type NullUpdateProtocol struct {
	UpdateProtocol UpdateProtocol
	Valid          bool // Valid is true if UpdateProtocol is not NULL
}

// Scan implements the Scanner interface.
func (ns *NullUpdateProtocol) Scan(value interface{}) error {
	if value == nil {
		ns.UpdateProtocol, ns.Valid = "", false
		return nil
	}
	ns.Valid = true
	return ns.UpdateProtocol.Scan(value)
}

// Value implements the driver Valuer interface.
func (ns NullUpdateProtocol) Value() (driver.Value, error) {
	if !ns.Valid {
		return nil, nil
	}
	return string(ns.UpdateProtocol), nil
}

type UpdateStatus string

const (
	UpdateStatusEmpty      UpdateStatus = "empty"
	UpdateStatusPending    UpdateStatus = "pending"
	UpdateStatusProcessing UpdateStatus = "processing"
	UpdateStatusPublished  UpdateStatus = "published"
	UpdateStatusFailed     UpdateStatus = "failed"
	UpdateStatusCanceled   UpdateStatus = "canceled"
)

func (e *UpdateStatus) Scan(src interface{}) error {
	switch s := src.(type) {
	case []byte:
		*e = UpdateStatus(s)
	case string:
		*e = UpdateStatus(s)
	default:
		return fmt.Errorf("unsupported scan type for UpdateStatus: %T", src)
	}
	return nil
}

type NullUpdateStatus struct {
	UpdateStatus UpdateStatus
	Valid        bool // Valid is true if UpdateStatus is not NULL
}

// Scan implements the Scanner interface.
func (ns *NullUpdateStatus) Scan(value interface{}) error {
	if value == nil {
		ns.UpdateStatus, ns.Valid = "", false
		return nil
	}
	ns.Valid = true
	return ns.UpdateStatus.Scan(value)
}

// Value implements the driver Valuer interface.
func (ns NullUpdateStatus) Value() (driver.Value, error) {
	if !ns.Valid {
		return nil, nil
	}
	return string(ns.UpdateStatus), nil
}

type Project struct {
	ID             uuid.UUID
	Name           string
	UpdateProtocol UpdateProtocol
	CreatedAt      pgtype.Timestamptz
}

type Update struct {
	ID             uuid.UUID
	ProjectID      uuid.UUID
	RuntimeVersion string
	Status         UpdateStatus
	Message        pgtype.Text
	Channel        string
	CreatedAt      pgtype.Timestamptz
}

type UpdateAsset struct {
	ID                uuid.UUID
	UpdateID          uuid.UUID
	StorageObjectPath string
	ContentType       string
	Extension         string
	ContentMd5        string
	ContentSha256     string
	IsLaunchAsset     bool
	IsArchive         bool
	Platform          string
	ContentLength     int64
	CreatedAt         pgtype.Timestamptz
}

type UpdateMetadatum struct {
	ID            uuid.UUID
	UpdateID      uuid.UUID
	ExpoAppConfig []byte
	CreatedAt     pgtype.Timestamptz
}